Make Lombok use consistent on all domain and dto classes. Move Pathways from IMMSCompound
