+++ src/main/java/ceu/biolab/cmm/ccsSearch/domain/AnnotationsByAdduct.java
@@ -3,6 +3,9 @@ package ceu.biolab.cmm.ccsSearch.domain;
 import java.util.List;
 import java.util.ArrayList;
 
+import lombok.Data;
+
+@Data
 public class AnnotationsByAdduct {
     private String adduct;
     private List<IMMSCompound> compounds;
@@ -21,28 +24,11 @@ public class AnnotationsByAdduct {
         this.compounds = new ArrayList<>();
     }
 
-    public String getAdduct() {
-        return adduct;
-    }
-
-    public void setAdduct(String adduct) {
-        this.adduct = adduct;
-    }
-
-    public List<IMMSCompound> getCompounds() {
-        return compounds;
-    }
-
     public void setCompounds(List<IMMSCompound> compounds) {
         if (compounds == null) {
             this.compounds = new ArrayList<>();
         } else {
             this.compounds = new ArrayList<>(compounds);
         }
-    }
-
-    @Override
-    public String toString() {
-        return "AnnotationsByAdduct [adduct=" + adduct + ", compounds=" + compounds + "]";
     }
 }
