+++ src/main/java/ceu/biolab/cmm/ccsSearch/domain/IMMSCompound.java
@@ -1,10 +1,6 @@
 package ceu.biolab.cmm.ccsSearch.domain;
 
-import java.util.HashSet;
-import java.util.Set;
-
 import ceu.biolab.cmm.shared.domain.compound.Compound;
-import ceu.biolab.cmm.shared.domain.compound.Pathway;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
@@ -14,8 +10,6 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 public class IMMSCompound extends Compound {
     private double dbCcs;
-    // TODO pathways should be on the shared domain
-    private Set<Pathway> pathways;
 
     public IMMSCompound(Compound compound, double ccsValue) {
         super(compound.getCompoundId(), compound.getCasId(), compound.getCompoundName(), compound.getFormula(),
@@ -26,15 +20,5 @@ public class IMMSCompound extends Compound {
               compound.getDoubleBonds(), compound.getBiologicalActivity(), compound.getMeshNomenclature(),
               compound.getIupacClassification(), compound.getMol2(), compound.getPathways());
         this.dbCcs = ccsValue;
-        this.pathways = new HashSet<>();
-    }
-
-    public void addPathway(Pathway pathway) {
-        if (this.pathways == null) {
-            this.pathways = new HashSet<>();
-        }
-        if (pathway != null && pathway.getPathwayId() != -1) {
-            this.pathways.add(pathway);
-        }
     }
 }
